Implement new approach for the knight moves
The previous dabblings have mostly been removed. Just the interactions and board building are left.

Currently trying to implement a different solution using a MOVES variable listing all the possible knight moves and then using this to generate all the possible valid squares the knight can go to. These squares will be stored as an instance variable in each Knight.

From the most probably go for a BFS method, generating new Knights as needed on the go.