Build a tree with possible Knight moves
The tree will go 6 levels deep at most, as this is the maximum amount of moves you can need to reach any square. It doesn't filter out double positions just yet, as I didn't find a way to ensure the path to the position would always be shortest.

This might be done by first generating all possible moves in a cartain level, putting these in an array, then continue working on the next moves. Maybe to be added later.

Next up will be to implement a search algorithm that will search the full level first before moving on to the next one to ensure shortest path.